+++ src/Entity/WorkComment.php
@@ -3,6 +3,7 @@
 namespace App\Entity;
 
 use App\Repository\WorkCommentRepository;
+use App\Trait\CommentFieldsTrait;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
@@ -14,10 +15,7 @@ use Symfony\Component\Validator\Constraints as Assert;
 #[ORM\Index(columns: ['parent_id'], name: 'idx_work_comment_parent')]
 class WorkComment
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column(type: Types::INTEGER)]
-    private ?int $id = null;
+    use CommentFieldsTrait;
 
     #[ORM\ManyToOne(targetEntity: Work::class, inversedBy: 'comments')]
     #[ORM\JoinColumn(name: 'work_id', referencedColumnName: 'id', nullable: false, onDelete: 'CASCADE')]
@@ -31,39 +29,10 @@ class WorkComment
     #[ORM\OrderBy(['createdAt' => 'ASC'])]
     private Collection $children;
 
-    #[ORM\Column(type: Types::STRING, length: 100)]
-    #[Assert\NotBlank]
-    #[Assert\Length(max: 100)]
-    private string $author = '';
-
-    #[ORM\Column(type: Types::TEXT)]
-    #[Assert\NotBlank]
-    private string $content = '';
-
-    #[ORM\Column(type: Types::STRING, length: 255, nullable: true)]
-    private ?string $info = null;
-
-    #[ORM\Column(type: Types::DATETIME_IMMUTABLE)]
-    private \DateTimeImmutable $createdAt;
-
-    #[ORM\Column(type: Types::BOOLEAN)]
-    private bool $isActive = true;
-
     public function __construct()
     {
         $this->children = new ArrayCollection();
         $this->createdAt = new \DateTimeImmutable();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function setId(int $id): static
-    {
-        $this->id = $id;
-        return $this;
     }
 
     public function getWork(): ?Work
@@ -99,6 +68,7 @@ class WorkComment
             $this->children->add($child);
             $child->setParent($this);
         }
+
         return $this;
     }
 
@@ -109,71 +79,7 @@ class WorkComment
                 $child->setParent(null);
             }
         }
-        return $this;
-    }
 
-    public function __toString(): string
-    {
-        $content = strip_tags($this->content);
-        return strip_tags(
-            mb_strlen($content) > 150
-                ? mb_substr($content, 0, 147) . '…'
-                : $content
-        );
-    }
-
-    public function getAuthor(): string
-    {
-        return $this->author;
-    }
-
-    public function setAuthor(string $author): static
-    {
-        $this->author = $author;
-        return $this;
-    }
-
-    public function getContent(): string
-    {
-        return $this->content;
-    }
-
-    public function setContent(string $content): static
-    {
-        $this->content = $content;
-        return $this;
-    }
-
-    public function getInfo(): ?string
-    {
-        return $this->info;
-    }
-
-    public function setInfo(?string $info): static
-    {
-        $this->info = $info;
-        return $this;
-    }
-
-    public function getCreatedAt(): \DateTimeImmutable
-    {
-        return $this->createdAt;
-    }
-
-    public function setCreatedAt(\DateTimeImmutable $createdAt): static
-    {
-        $this->createdAt = $createdAt;
-        return $this;
-    }
-
-    public function getIsActive(): bool
-    {
-        return $this->isActive;
-    }
-
-    public function setIsActive(bool $isActive): static
-    {
-        $this->isActive = $isActive;
         return $this;
     }
 }