+++ src/Trait/HasDefaultTitleTrait.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Trait;
+
+trait HasDefaultTitleTrait
+{
+    abstract protected function getBodyContent(): string;
+
+    abstract protected function getDefaultComment();
+
+    public function getDefaultTitle(): string
+    {
+        $lines = explode("\n", trim($this->getBodyContent()));
+        $firstLine = trim(rtrim($lines[0], '.,!:?;…—-'));
+
+        return '"' . $firstLine . '..."';
+    }
+
+    public function getDisplayTitle(): string
+    {
+        return preg_match('/\".*\.\.\.\"$/', $this->getTitle())
+            ? '* * *'
+            : mb_strtoupper($this->getTitle() ?? '');
+    }
+
+    private function ensureDefaults(): void
+    {
+        if (empty($this->getTitle())) {
+            $this->setTitle($this->getDefaultTitle());
+        }
+        if (empty($this->getComment())) {
+            $this->setComment($this->getDefaultComment());
+        }
+    }
+}